TEDEFO-4980 Remove cross-notice reference support from V2 translator
